test_gw_repeater: disable initial operational test
https://jira.prplfoundation.org/browse/PPM-143

Disable operational test for environmental docker launch.

In certification mode we want to change agents logic to wait for
dev_set_config or WPS PBC in order to proceed to master discovery.

In environmental docker launch subroutine the check for operational state
is done right after launching the containers.

One of the options is changing `test_gw_repeater.sh` in order to send
dev_set_config there, but for sake of simplicity it was chosen to send
dev_set_config in `test_flows.py` (will be introduced in a later commit in this PR).

There is also an undesirable effect of reporting repetears as unoperational
even though they will be usable after sending dev_set_config.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>