agent: Move 'controller_bridge_mac' to Agent DB
Move 'controller_bridge_mac' to Agent DB to Agent DB and use it instead of
local class members on the backhaul manager and the son_slave.

On the backhaul manager autoconfiguration_response handler, the controller
bridge mac was set twice, one time happens always, and the other time when
there is a match between the radio band type to the one received on the
response message. The second set of this parameter is redundant and
therefore removed.

The son_slave use to set the controller bridge mac on the master_socket with
"setPeerMac". This is redundant since 'getPeerMac' is not used anywhere on
the son_slave, therefore removed.

PPM-345

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>